chore: rename /models/fight.ts to fight-model.ts
