+++ models/fight-model.ts
